Improve seticons function

+++ .config/fish/functions/seticons.fish
@@ -1,5 +1,6 @@
 #!/usr/bin/env fish
 function seticons
+    set haschanged 0
     for file in ~/icons/*
         if not test -e "$file"
             continue
@@ -12,8 +13,16 @@ function seticons
         set newname (string replace -r '\.[^.]+$' '.app' -- $basename)
 
         set appname "/Applications/$newname"
-
-        sudo fileicon set $appname $file
+    
+        if not fileicon test $appname
+            sudo fileicon set $appname $file
+            set haschanged 1
+        end
     end
-    killall Dock
+    if test $haschanged -eq 1
+        echo restarting dock
+        rm /var/folders/*/*/*/com.apple.dock.iconcache
+        rm -r /var/folders/*/*/*/com.apple.iconservices*
+        killall Dock
+    end
 end
